Refactor folder sturcture
